+++ src/mefit/Program.cs
@@ -41,8 +41,8 @@ namespace Mac_EFI_Toolkit
     internal readonly struct METVersion
     {
         internal const string LZMA_SDK = "24.08";
-        internal const string APP_BUILD = "241113.0330";
-        internal const string APP_CHANNEL = "BETA 3";
+        internal const string APP_BUILD = "241114.0045";
+        internal const string APP_CHANNEL = "BETA 4";
     }
 
     internal readonly struct METUrl
@@ -79,6 +79,7 @@ namespace Mac_EFI_Toolkit
         internal const string GLYPH_RIGHT_ARROW = "\u2192";
         internal const string GLYPH_VIEW = "\uE890";
         internal const string GLYPH_ZIP = "\uF012";
+        internal const string GLYPH_USER = "\uE77B";
         internal const string NOWRAP_SPACE = "\u00A0";
 
         internal static Font FONT_MDL2_REG_9;